Add troubleshooting for windows 11 x86 guest on arm host
As per #4804 the "This PC can't run Windows 11" error troubleshooting guide isn't correct, at least in the case of windows x86 guest on arm host.
I found that the TPM is on by default when using the VM creation assistant, but the default number of threads seems to be 1 (although not shown). And this seems to be the problem for windows 11 installation, it prevents install with the error. Setting the cpu core count to 2 or more fixes the issue. 
I think the default should changed to 2 instead of 1. In the meantime I suggest adding this info to the troubleshooting guide